pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pavgls-experiment"
version = "0.1.0"
description = "Phase-Adaptive Variance-Guided Learning Rate Scheduling (PAVGLS) Experiment"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "wandb>=0.15.0",
    "numpy>=1.24.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "optuna>=3.0.0",
    "scipy>=1.10.0",
    "pandas>=2.0.0",
    "pyyaml>=6.0",
]
